+++ src/app/api/auth/[...nextauth]/options.js
@@ -1,0 +1,48 @@
+import GitHubProvider from "next-auth/providers/github";
+import GoogleProvider from "next-auth/providers/google";
+import FacebookProvider from "next-auth/providers/facebook";
+
+export const options = {
+  providers: [
+   FacebookProvider({
+        profile(profile) {
+        console.log("Facebook user: ", profile);
+
+        let userRole = "Facebook User"
+        return {
+          ...profile,
+          id: profile.sub,
+          role: userRole,
+        };
+      },
+      clientId: process.env.FACEBOOK_CLIENT_ID,
+      clientSecret: process.env.FACEBOOK_CLIENT_SECRET,
+    }),
+
+    GoogleProvider({
+      profile(profile) {
+        console.log("Profile Google: ", profile);
+
+        let userRole = "Google User"
+        return {
+          ...profile,
+          id: profile.sub,
+          role: userRole,
+        };
+      },
+      clientId: process.env.GOOGLE_ID,
+      clientSecret: process.env.GOOGLE_Secret,
+    }),
+  ],
+
+  callbacks: {
+    async jwt({ token, user }) {
+      if (user) token.role = user.role;
+      return token;
+    },
+    async session({ session, token }) {
+      if (session?.user) session.user.role = token.role;
+      return session;
+    },
+  },
+};
