+++ src/app/api/auth/[...nextauth]/route.js
@@ -1,0 +1,6 @@
+import NextAuth from "next-auth";
+
+import { options } from "./options"
+
+const handler = NextAuth(options);
+export {handler as GET, handler as POST}
