+++ src/components/AuthProvider.js
@@ -1,0 +1,9 @@
+"use client"
+
+import { SessionProvider } from "next-auth/react"
+
+const AuthProvider = ({children}) => {
+    return <SessionProvider> {children} </SessionProvider>
+}
+
+export default AuthProvider;
